reuse read(byte[] b) which is compatible for reading bytes from an AudioInputStream on Java 8
The error you're encountering, java.lang.NoSuchMethodError: javax.sound.sampled.AudioInputStream.readAllBytes(), indicates that the method readAllBytes() does not exist in the AudioInputStream class. This is because the readAllBytes() method was introduced in Java 9, but AWS Lambda typically runs on an older version of Java (Java 8), where this method is not available.